requirements.txt
annotated-types==0.7.0
anyio==4.6.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.5
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
